fix: parser ui: storage layout changed
