+++ bootstrap/helpers/proxy.php
@@ -151,7 +151,7 @@ function generate_default_proxy_configuration(Server $server)
             'services' => [
                 'traefik' => [
                     'container_name' => 'coolify-proxy',
-                    'image' => 'traefik:v2.11',
+                    'image' => 'traefik:v3.1',
                     'restart' => RESTART_MODE,
                     'extra_hosts' => [
                         'host.docker.internal:host-gateway',
