+++ bootstrap/helpers/shared.php
@@ -1,19 +1,15 @@
 <?php
 
 use App\Enums\ApplicationDeploymentStatus;
-use App\Enums\ProxyTypes;
 use App\Jobs\ServerFilesFromServerJob;
 use App\Models\Application;
 use App\Models\ApplicationDeploymentQueue;
-use App\Models\ApplicationPreview;
-use App\Models\EnvironmentVariable;
 use App\Models\InstanceSettings;
 use App\Models\LocalFileVolume;
 use App\Models\LocalPersistentVolume;
 use App\Models\Server;
 use App\Models\Service;
 use App\Models\ServiceApplication;
-use App\Models\ServiceDatabase;
 use App\Models\StandaloneClickhouse;
 use App\Models\StandaloneDragonfly;
 use App\Models\StandaloneKeydb;
@@ -778,539 +774,11 @@ function replaceLocalSource(Stringable $source, Stringable $replacedWith)
     if ($source->startsWith('..')) {
         $source = $source->replaceFirst('..', $replacedWith->value());
     }
+    if ($source->endsWith('/')) {
+        $source = $source->replaceLast('/', '');
+    }
 
     return $source;
-}
-function dockerComposeParserForApplications(Application $application): Collection
-{
-    $pullRequestId = data_get($application, 'pull_request_id', 0);
-    $isPullRequest = $pullRequestId === 0 ? false : true;
-
-    $uuid = data_get($application, 'uuid');
-    $server = data_get($application, 'destination.server');
-    $compose = data_get($application, 'docker_compose_raw');
-    $yaml = Yaml::parse($compose);
-    $services = data_get($yaml, 'services', collect([]));
-    $topLevel = collect([
-        'volumes' => collect(data_get($compose, 'volumes', [])),
-        'networks' => collect(data_get($compose, 'networks', [])),
-        'configs' => collect(data_get($compose, 'configs', [])),
-        'secrets' => collect(data_get($compose, 'secrets', [])),
-    ]);
-
-    // If there are predefined volumes, make sure they are not null
-    if ($topLevel->get('volumes')->count() > 0) {
-        $temp = collect([]);
-        foreach ($topLevel['volumes'] as $volumeName => $volume) {
-            if (is_null($volume)) {
-                continue;
-            }
-            $temp->put($volumeName, $volume);
-        }
-        $topLevel['volumes'] = $temp;
-    }
-    // Get the base docker network
-    $baseNetwork = collect([$uuid]);
-    if ($isPullRequest) {
-        $baseNetwork = collect(["{$uuid}-{$pullRequestId}"]);
-    }
-    $parsedServices = collect([]);
-    $fileStorages = $application->fileStorages();
-    // Let's loop through the services
-    foreach ($services as $serviceName => $service) {
-        $isDatabase = isDatabaseImage(data_get_str($service, 'image'));
-        $image = data_get_str($service, 'image');
-        $restart = data_get_str($service, 'restart', RESTART_MODE);
-        $logging = data_get($service, 'logging');
-        $healthcheck = data_get($service, 'healthcheck');
-
-        if ($server->isLogDrainEnabled() && $application->isLogDrainEnabled()) {
-            $logging = [
-                'driver' => 'fluentd',
-                'options' => [
-                    'fluentd-address' => 'tcp://127.0.0.1:24224',
-                    'fluentd-async' => 'true',
-                    'fluentd-sub-second-precision' => 'true',
-                ],
-            ];
-        }
-
-        $volumes = collect(data_get($service, 'volumes', []));
-        $ports = collect(data_get($service, 'ports', []));
-        $networks = collect(data_get($service, 'networks', []));
-        $depends_on = collect(data_get($service, 'depends_on', []));
-        $labels = collect(data_get($service, 'labels', []));
-        $environment = collect(data_get($service, 'environment', []));
-        $buildArgs = collect(data_get($service, 'build.args', []));
-        $environment = $environment->merge($buildArgs);
-
-        // TODO: refactor generateApplicationContainerName
-        $baseName = generateApplicationContainerName(
-            application: $application,
-            pull_request_id: $pullRequestId
-        );
-        $containerName = "$serviceName-$baseName";
-        $volumesParsed = collect([]);
-
-        if ($volumes->count() > 0) {
-            foreach ($volumes as $index => $volume) {
-                $type = null;
-                $source = null;
-                $target = null;
-                $content = null;
-                $isDirectory = false;
-                if (is_string($volume)) {
-                    $source = str($volume)->before(':');
-                    $target = str($volume)->after(':')->beforeLast(':');
-                    $foundConfig = $fileStorages->whereMountPath($target)->first();
-                    if (sourceIsLocal($source)) {
-                        $type = str('bind');
-                        if ($foundConfig) {
-                            $contentNotNull_temp = data_get($foundConfig, 'content');
-                            if ($contentNotNull_temp) {
-                                $content = $contentNotNull_temp;
-                            }
-                            $isDirectory = data_get($foundConfig, 'is_directory');
-                        } else {
-                            // By default, we cannot determine if the bind is a directory or not, so we set it to directory
-                            $isDirectory = true;
-                        }
-                    } else {
-                        $type = str('volume');
-                    }
-                } elseif (is_array($volume)) {
-                    $type = data_get_str($volume, 'type');
-                    $source = data_get_str($volume, 'source');
-                    $target = data_get_str($volume, 'target');
-                    $content = data_get($volume, 'content');
-                    $isDirectory = (bool) data_get($volume, 'isDirectory', null) || (bool) data_get($volume, 'is_directory', null);
-
-                    $foundConfig = $fileStorages->whereMountPath($target)->first();
-                    if ($foundConfig) {
-                        $contentNotNull_temp = data_get($foundConfig, 'content');
-                        if ($contentNotNull_temp) {
-                            $content = $contentNotNull_temp;
-                        }
-                        $isDirectory = data_get($foundConfig, 'is_directory');
-                    } else {
-                        // if isDirectory is not set (or false) & content is also not set, we assume it is a directory
-                        if ((is_null($isDirectory) || ! $isDirectory) && is_null($content)) {
-                            $isDirectory = true;
-                        }
-                    }
-                }
-                if ($type->value() === 'bind') {
-                    if ($source->value() === '/var/run/docker.sock') {
-                        return $volume;
-                    }
-                    if ($source->value() === '/tmp' || $source->value() === '/tmp/') {
-                        return $volume;
-                    }
-                    $mainDirectory = str(base_configuration_dir().'/applications/'.$uuid);
-                    $source = replaceLocalSource($source, $mainDirectory);
-                    if ($isPullRequest) {
-                        $source = $source."-pr-$pullRequestId";
-                    }
-                    if (
-                        ! $application?->settings?->is_preserve_repository_enabled || $foundConfig?->is_based_on_git
-                    ) {
-                        // ray([
-                        //     'fs_path' => $source->value(),
-                        //     'mount_path' => $target->value(),
-                        //     'content' => $content,
-                        //     'is_directory' => $isDirectory,
-                        //     'resource_id' => $application->id,
-                        //     'resource_type' => get_class($application),
-                        // ]);
-                        // LocalFileVolume::updateOrCreate(
-                        //     [
-                        //         'mount_path' => $target,
-                        //         'resource_id' => $application->id,
-                        //         'resource_type' => get_class($application),
-                        //     ],
-                        //     [
-                        //         'fs_path' => $source,
-                        //         'mount_path' => $target,
-                        //         'content' => $content,
-                        //         'is_directory' => $isDirectory,
-                        //         'resource_id' => $application->id,
-                        //         'resource_type' => get_class($application),
-                        //     ]
-                        // );
-                    }
-                    $volume = "$source:$target";
-                } elseif ($type->value() === 'volume') {
-                    if ($topLevel->get('volumes')->has($source->value())) {
-                        $temp = $topLevel->get('volumes')->get($source->value());
-                        if (data_get($temp, 'driver_opts.type') === 'cifs') {
-                            return $volume;
-                        }
-                        if (data_get($temp, 'driver_opts.type') === 'nfs') {
-                            return $volume;
-                        }
-                    }
-                    $slugWithoutUuid = Str::slug($source, '-');
-                    $name = "{$uuid}_{$slugWithoutUuid}";
-                    if (is_string($volume)) {
-                        $source = str($volume)->before(':');
-                        $target = str($volume)->after(':')->beforeLast(':');
-                        $source = $name;
-                        $volume = "$source:$target";
-                    } elseif (is_array($volume)) {
-                        data_set($volume, 'source', $name);
-                    }
-                    $topLevel->get('volumes')->put($name, [
-                        'name' => $name,
-                    ]);
-                    // ray([
-                    //     'name' => $name,
-                    //     'mount_path' => $target,
-                    //     'resource_id' => $application->id,
-                    //     'resource_type' => get_class($application),
-                    // ]);
-                    // LocalPersistentVolume::updateOrCreate(
-                    //     [
-                    //         'mount_path' => $target,
-                    //         'resource_id' => $application->id,
-                    //         'resource_type' => get_class($application),
-                    //     ],
-                    //     [
-                    //         'name' => $name,
-                    //         'mount_path' => $target,
-                    //         'resource_id' => $application->id,
-                    //         'resource_type' => get_class($application),
-                    //     ]
-                    // );
-                }
-                // dispatch(new ServerFilesFromServerJob($application));
-                $volumesParsed->put($index, $volume);
-            }
-        }
-        if ($depends_on?->count() > 0) {
-            if ($isPullRequest) {
-                $depends_on->transform(function ($dependency) use ($pullRequestId) {
-                    return "$dependency-pr-$pullRequestId";
-                });
-            }
-        }
-
-        if ($topLevel->get('networks')?->count() > 0) {
-            foreach ($topLevel->get('networks') as $networkName => $network) {
-                if ($networkName === 'default') {
-                    continue;
-                }
-                // ignore aliases
-                if ($network['aliases'] ?? false) {
-                    continue;
-                }
-                $networkExists = $networks->contains(function ($value, $key) use ($networkName) {
-                    return $value == $networkName || $key == $networkName;
-                });
-                if (! $networkExists) {
-                    $networks->put($networkName, null);
-                }
-            }
-        }
-        $baseNetworkExists = $networks->contains(function ($value, $_) use ($baseNetwork) {
-            return $value == $baseNetwork;
-        });
-        if (! $baseNetworkExists) {
-            foreach ($baseNetwork as $network) {
-                $topLevel->get('networks')->put($network, [
-                    'name' => $network,
-                    'external' => true,
-                ]);
-            }
-        }
-        $networks_temp = collect();
-
-        foreach ($networks as $key => $network) {
-            if (gettype($network) === 'string') {
-                // networks:
-                //  - appwrite
-                $networks_temp->put($network, null);
-            } elseif (gettype($network) === 'array') {
-                // networks:
-                //   default:
-                //     ipv4_address: 192.168.203.254
-                $networks_temp->put($key, $network);
-            }
-        }
-        foreach ($baseNetwork as $key => $network) {
-            $networks_temp->put($network, null);
-        }
-
-        if (data_get($application, 'settings.connect_to_docker_network')) {
-            $network = $application->destination->network;
-            $networks_temp->put($network, null);
-            $topLevel->get('networks')->put($network, [
-                'name' => $network,
-                'external' => true,
-            ]);
-        }
-
-        foreach ($environment as $key => $value) {
-            if (is_numeric($key)) {
-                if (is_array($value)) {
-                    // - SESSION_SECRET: 123
-                    // - SESSION_SECRET:
-                    $key = str(collect($value)->keys()->first());
-                    $value = str(collect($value)->values()->first());
-                } else {
-                    $variable = str($value);
-                    if ($variable->contains('=')) {
-                        // - SESSION_SECRET=123
-                        // - SESSION_SECRET=
-                        $key = $variable->before('=');
-                        $value = $variable->after('=');
-                    } else {
-                        // - SESSION_SECRET
-                        $key = $variable;
-                        $value = null;
-                    }
-                }
-            } else {
-                // SESSION_SECRET: 123
-                // SESSION_SECRET:
-                $key = str($key);
-                $value = str($value);
-            }
-
-            // Auto generate FQDN and URL
-            if ($key->startsWith('SERVICE_FQDN') || $value->startsWith('$SERVICE_FQDN') || $value->startsWith('${SERVICE_FQDN')) {
-                if ($value->contains('SERVICE_FQDN')) {
-                    $key = str(replaceVariables($value));
-                    $value = null;
-                }
-                $name = $key->after('SERVICE_FQDN_')->beforeLast('_')->lower();
-                $fqdn = generateFqdn($server, "{$name->value()}-{$uuid}");
-
-                $url = str($fqdn)->replace('http://', '')->replace('https://', '')->replace('www.', '');
-                $keyUrl = $key->replace('SERVICE_FQDN', 'SERVICE_URL');
-                // TODO: is this needed?
-                // if (substr_count($key->value(), '_') === 3) {
-                //     // SERVICE_FQDN_UMAMI_1000
-                //     $port = $key->afterLast('_');
-                // } else {
-                //     // SERVICE_FQDN_UMAMI
-                //     $port = null;
-                // }
-                // if ($port) {
-                //     $fqdn = "$fqdn:$port";
-                // }
-                if ($value && get_class($value) === 'Illuminate\Support\Stringable') {
-                    $path = $value->value();
-                    $fqdn = "$fqdn$path";
-                }
-
-                // ray([
-                //     'key' => $key,
-                //     'value' => $fqdn,
-                // ]);
-                // ray($application->environment_variables()->where('key', $key)->where('application_id', $application->id)->first());
-                $application->environment_variables()->where('key', $key)->where('application_id', $application->id)->firstOrCreate([
-                    'key' => $key,
-                    'application_id' => $application->id,
-                ], [
-                    'value' => $fqdn,
-                    'is_build_time' => false,
-                    'is_preview' => false,
-                ]);
-                $application->environment_variables()->where('key', $keyUrl)->where('application_id', $application->id)->firstOrCreate([
-                    'key' => $keyUrl,
-                    'application_id' => $application->id,
-                ], [
-                    'value' => $url,
-                    'is_build_time' => false,
-                    'is_preview' => false,
-                ]);
-            } elseif ($value->startsWith('$')) {
-                // If the value is a variable then we will add it to Coolify's DB
-
-                // ${VARIABLE} will be VARIABLE instead
-                $value = str(replaceVariables($value));
-                if ($value->contains(':-')) {
-                    $key = $value->before(':');
-                    $defaultValue = $value->after(':-');
-                } elseif ($value->contains('-')) {
-                    $key = $value->before('-');
-                    $defaultValue = $value->after('-');
-                } elseif ($value->contains(':?')) {
-                    $key = $value->before(':');
-                    $defaultValue = $value->after(':?');
-                } elseif ($value->contains('?')) {
-                    $key = $value->before('?');
-                    $defaultValue = $value->after('?');
-                } else {
-                    $key = $value;
-                    $defaultValue = null;
-                }
-                $application->environment_variables()->where('key', $key)->where('application_id', $application->id)->firstOrCreate([
-                    'key' => $key,
-                    'application_id' => $application->id,
-                    'is_preview' => false,
-                ], [
-                    'value' => $defaultValue,
-                    'is_build_time' => false,
-                    'is_preview' => false,
-                ]);
-            }
-        }
-
-        // Labels
-        $fqdns = collect([]);
-        if ($application?->serviceType()) {
-            $fqdns = generateServiceSpecificFqdns($application);
-        } else {
-            $domains = collect(json_decode($application->docker_compose_domains)) ?? collect([]);
-            if ($domains->count() !== 0) {
-                $fqdns = data_get($domains, "$serviceName.domain");
-                if (! $fqdns) {
-                    $fqdns = collect([]);
-                } else {
-                    $fqdns = str($fqdns)->explode(',');
-                    if ($isPullRequest) {
-                        $preview = $application->previews()->find($pullRequestId);
-                        $docker_compose_domains = collect(json_decode(data_get($preview, 'docker_compose_domains')));
-                        if ($docker_compose_domains->count() > 0) {
-                            $found_fqdn = data_get($docker_compose_domains, "$serviceName.domain");
-                            if ($found_fqdn) {
-                                $fqdns = collect($found_fqdn);
-                            } else {
-                                $fqdns = collect([]);
-                            }
-                        } else {
-                            $fqdns = $fqdns->map(function ($fqdn) use ($pullRequestId) {
-                                $preview = ApplicationPreview::findPreviewByApplicationAndPullId($this->id, $pullRequestId);
-                                $url = Url::fromString($fqdn);
-                                $template = $this->preview_url_template;
-                                $host = $url->getHost();
-                                $schema = $url->getScheme();
-                                $random = new Cuid2;
-                                $preview_fqdn = str_replace('{{random}}', $random, $template);
-                                $preview_fqdn = str_replace('{{domain}}', $host, $preview_fqdn);
-                                $preview_fqdn = str_replace('{{pr_id}}', $pullRequestId, $preview_fqdn);
-                                $preview_fqdn = "$schema://$preview_fqdn";
-                                $preview->fqdn = $preview_fqdn;
-                                $preview->save();
-
-                                return $preview_fqdn;
-                            });
-                        }
-                    }
-                    $shouldGenerateLabelsExactly = $server->settings->generate_exact_labels;
-                    if ($shouldGenerateLabelsExactly) {
-                        switch ($server->proxyType()) {
-                            case ProxyTypes::TRAEFIK->value:
-                                $labels = $labels->merge(
-                                    fqdnLabelsForTraefik(
-                                        uuid: $application->uuid,
-                                        domains: $fqdns,
-                                        serviceLabels: $labels,
-                                        generate_unique_uuid: $application->build_pack === 'dockercompose',
-                                        image: $image,
-                                        is_force_https_enabled: $application->isForceHttpsEnabled(),
-                                        is_gzip_enabled: $application->isGzipEnabled(),
-                                        is_stripprefix_enabled: $application->isStripprefixEnabled(),
-                                    )
-                                );
-                                break;
-                            case ProxyTypes::CADDY->value:
-                                $labels = $labels->merge(
-                                    fqdnLabelsForCaddy(
-                                        network: $application->destination->network,
-                                        uuid: $application->uuid,
-                                        domains: $fqdns,
-                                        serviceLabels: $labels,
-                                        image: $image,
-                                        is_force_https_enabled: $application->isForceHttpsEnabled(),
-                                        is_gzip_enabled: $application->isGzipEnabled(),
-                                        is_stripprefix_enabled: $application->isStripprefixEnabled(),
-                                    )
-                                );
-                                break;
-                        }
-                    } else {
-                        $labels = $labels->merge(
-                            fqdnLabelsForTraefik(
-                                uuid: $application->uuid,
-                                domains: $fqdns,
-                                serviceLabels: $labels,
-                                generate_unique_uuid: $application->build_pack === 'dockercompose',
-                                image: $image,
-                                is_force_https_enabled: $application->isForceHttpsEnabled(),
-                                is_gzip_enabled: $application->isGzipEnabled(),
-                                is_stripprefix_enabled: $application->isStripprefixEnabled(),
-                            )
-                        );
-                        $labels = $labels->merge(
-                            fqdnLabelsForCaddy(
-                                network: $application->destination->network,
-                                uuid: $application->uuid,
-                                domains: $fqdns,
-                                serviceLabels: $labels,
-                                image: $image,
-                                is_force_https_enabled: $application->isForceHttpsEnabled(),
-                                is_gzip_enabled: $application->isGzipEnabled(),
-                                is_stripprefix_enabled: $application->isStripprefixEnabled(),
-                            )
-                        );
-                    }
-                }
-            }
-        }
-
-        $defaultLabels = defaultLabels(
-            id: $application->id,
-            name: $containerName,
-            pull_request_id: $pullRequestId,
-            type: 'application');
-        $labels = $labels->merge($defaultLabels);
-
-        if ($labels->count() > 0 && $application->settings->is_container_label_escape_enabled) {
-            $labels = $labels->map(function ($value, $key) {
-                return escapeDollarSign($value);
-            });
-        }
-        $payload = [
-            'image' => $image,
-            'restart' => $restart,
-            'container_name' => $containerName,
-            'volumes' => $volumesParsed,
-            'networks' => $networks_temp,
-            'labels' => $labels,
-            'environment' => $environment,
-
-        ];
-        if ($ports->count() > 0) {
-            $payload['ports'] = $ports;
-        }
-        if ($logging) {
-            $payload['logging'] = $logging;
-        }
-        if ($depends_on->count() > 0) {
-            $payload['depends_on'] = $depends_on;
-        }
-        if ($healthcheck) {
-            $payload['healthcheck'] = $healthcheck;
-        }
-
-        $parsedServices->put($serviceName, $payload);
-
-    }
-
-    $topLevel->put('services', $parsedServices);
-    $customOrder = ['services', 'volumes', 'networks', 'configs', 'secrets'];
-
-    $topLevel = $topLevel->sortBy(function ($value, $key) use ($customOrder) {
-        return array_search($key, $customOrder);
-    });
-    $application->docker_compose = Yaml::dump(convertToArray($topLevel), 10, 2);
-    data_forget($application, 'environment_variables');
-    data_forget($application, 'environment_variables_preview');
-    $application->save();
-
-    return $topLevel;
 }
 
 function convertToArray($collection)
@@ -1328,602 +796,6 @@ function convertToArray($collection)
     }
 
     return $collection;
-}
-function parseDockerComposeFile(Service|Application $resource, bool $isNew = false, int $pull_request_id = 0, ?int $preview_id = null)
-{
-    if ($resource->getMorphClass() === 'App\Models\Service') {
-        if ($resource->docker_compose_raw) {
-            try {
-                $yaml = Yaml::parse($resource->docker_compose_raw);
-            } catch (\Exception $e) {
-                throw new \Exception($e->getMessage());
-            }
-            $allServices = get_service_templates();
-            $topLevelVolumes = collect(data_get($yaml, 'volumes', []));
-            $topLevelNetworks = collect(data_get($yaml, 'networks', []));
-            $topLevelConfigs = collect(data_get($yaml, 'configs', []));
-            $topLevelSecrets = collect(data_get($yaml, 'secrets', []));
-            $services = data_get($yaml, 'services');
-
-            $generatedServiceFQDNS = collect([]);
-            if (is_null($resource->destination)) {
-                $destination = $resource->server->destinations()->first();
-                if ($destination) {
-                    $resource->destination()->associate($destination);
-                    $resource->save();
-                }
-            }
-            $definedNetwork = collect([$resource->uuid]);
-            if ($topLevelVolumes->count() > 0) {
-                $tempTopLevelVolumes = collect([]);
-                foreach ($topLevelVolumes as $volumeName => $volume) {
-                    if (is_null($volume)) {
-                        continue;
-                    }
-                    $tempTopLevelVolumes->put($volumeName, $volume);
-                }
-                $topLevelVolumes = collect($tempTopLevelVolumes);
-            }
-            $services = collect($services)->map(function ($service, $serviceName) use ($topLevelNetworks, $definedNetwork, $isNew, $generatedServiceFQDNS, $resource, $allServices, $topLevelVolumes) {
-                // Workarounds for beta users.
-                if ($serviceName === 'registry') {
-                    $tempServiceName = 'docker-registry';
-                } else {
-                    $tempServiceName = $serviceName;
-                }
-                if (str(data_get($service, 'image'))->contains('glitchtip')) {
-                    $tempServiceName = 'glitchtip';
-                }
-                if ($serviceName === 'supabase-kong') {
-                    $tempServiceName = 'supabase';
-                }
-                $serviceDefinition = data_get($allServices, $tempServiceName);
-                $predefinedPort = data_get($serviceDefinition, 'port');
-                if ($serviceName === 'plausible') {
-                    $predefinedPort = '8000';
-                }
-                // End of workarounds for beta users.
-                $serviceVolumes = collect(data_get($service, 'volumes', []));
-                $servicePorts = collect(data_get($service, 'ports', []));
-                $serviceNetworks = collect(data_get($service, 'networks', []));
-                $serviceVariables = collect(data_get($service, 'environment', []));
-                $serviceLabels = collect(data_get($service, 'labels', []));
-                $hasHostNetworkMode = data_get($service, 'network_mode') === 'host' ? true : false;
-                if ($serviceLabels->count() > 0) {
-                    $removedLabels = collect([]);
-                    $serviceLabels = $serviceLabels->filter(function ($serviceLabel, $serviceLabelName) use ($removedLabels) {
-                        if (! str($serviceLabel)->contains('=')) {
-                            $removedLabels->put($serviceLabelName, $serviceLabel);
-
-                            return false;
-                        }
-
-                        return $serviceLabel;
-                    });
-                    foreach ($removedLabels as $removedLabelName => $removedLabel) {
-                        $serviceLabels->push("$removedLabelName=$removedLabel");
-                    }
-                }
-
-                $containerName = "$serviceName-{$resource->uuid}";
-
-                // Decide if the service is a database
-                $isDatabase = isDatabaseImage(data_get_str($service, 'image'));
-                $image = data_get_str($service, 'image');
-                data_set($service, 'is_database', $isDatabase);
-
-                // Create new serviceApplication or serviceDatabase
-                if ($isDatabase) {
-                    if ($isNew) {
-                        $savedService = ServiceDatabase::create([
-                            'name' => $serviceName,
-                            'image' => $image,
-                            'service_id' => $resource->id,
-                        ]);
-                    } else {
-                        $savedService = ServiceDatabase::where([
-                            'name' => $serviceName,
-                            'service_id' => $resource->id,
-                        ])->first();
-                    }
-                } else {
-                    if ($isNew) {
-                        $savedService = ServiceApplication::create([
-                            'name' => $serviceName,
-                            'image' => $image,
-                            'service_id' => $resource->id,
-                        ]);
-                    } else {
-                        $savedService = ServiceApplication::where([
-                            'name' => $serviceName,
-                            'service_id' => $resource->id,
-                        ])->first();
-                    }
-                }
-                if (is_null($savedService)) {
-                    if ($isDatabase) {
-                        $savedService = ServiceDatabase::create([
-                            'name' => $serviceName,
-                            'image' => $image,
-                            'service_id' => $resource->id,
-                        ]);
-                    } else {
-                        $savedService = ServiceApplication::create([
-                            'name' => $serviceName,
-                            'image' => $image,
-                            'service_id' => $resource->id,
-                        ]);
-                    }
-                }
-
-                // Check if image changed
-                if ($savedService->image !== $image) {
-                    $savedService->image = $image;
-                    $savedService->save();
-                }
-                // Collect/create/update networks
-                if ($serviceNetworks->count() > 0) {
-                    foreach ($serviceNetworks as $networkName => $networkDetails) {
-                        if ($networkName === 'default') {
-                            continue;
-                        }
-                        // ignore alias
-                        if ($networkDetails['aliases'] ?? false) {
-                            continue;
-                        }
-                        $networkExists = $topLevelNetworks->contains(function ($value, $key) use ($networkName) {
-                            return $value == $networkName || $key == $networkName;
-                        });
-                        if (! $networkExists) {
-                            $topLevelNetworks->put($networkDetails, null);
-                        }
-                    }
-                }
-
-                // Collect/create/update ports
-                $collectedPorts = collect([]);
-                if ($servicePorts->count() > 0) {
-                    foreach ($servicePorts as $sport) {
-                        if (is_string($sport) || is_numeric($sport)) {
-                            $collectedPorts->push($sport);
-                        }
-                        if (is_array($sport)) {
-                            $target = data_get($sport, 'target');
-                            $published = data_get($sport, 'published');
-                            $protocol = data_get($sport, 'protocol');
-                            $collectedPorts->push("$target:$published/$protocol");
-                        }
-                    }
-                }
-                $savedService->ports = $collectedPorts->implode(',');
-                $savedService->save();
-
-                if (! $hasHostNetworkMode) {
-                    // Add Coolify specific networks
-                    $definedNetworkExists = $topLevelNetworks->contains(function ($value, $_) use ($definedNetwork) {
-                        return $value == $definedNetwork;
-                    });
-                    if (! $definedNetworkExists) {
-                        foreach ($definedNetwork as $network) {
-                            $topLevelNetworks->put($network, [
-                                'name' => $network,
-                                'external' => true,
-                            ]);
-                        }
-                    }
-                    $networks = collect();
-                    foreach ($serviceNetworks as $key => $serviceNetwork) {
-                        if (gettype($serviceNetwork) === 'string') {
-                            // networks:
-                            //  - appwrite
-                            $networks->put($serviceNetwork, null);
-                        } elseif (gettype($serviceNetwork) === 'array') {
-                            // networks:
-                            //   default:
-                            //     ipv4_address: 192.168.203.254
-                            // $networks->put($serviceNetwork, null);
-                            $networks->put($key, $serviceNetwork);
-                        }
-                    }
-                    foreach ($definedNetwork as $key => $network) {
-                        $networks->put($network, null);
-                    }
-                    data_set($service, 'networks', $networks->toArray());
-                }
-
-                // Collect/create/update volumes
-                if ($serviceVolumes->count() > 0) {
-                    ['serviceVolumes' => $serviceVolumes, 'topLevelVolumes' => $topLevelVolumes] = parseServiceVolumes($serviceVolumes, $savedService, $topLevelVolumes);
-                    data_set($service, 'volumes', $serviceVolumes->toArray());
-                }
-
-                // Get variables from the service
-                foreach ($serviceVariables as $variableName => $variable) {
-                    if (is_numeric($variableName)) {
-                        if (is_array($variable)) {
-                            // - SESSION_SECRET: 123
-                            // - SESSION_SECRET:
-                            $key = str(collect($variable)->keys()->first());
-                            $value = str(collect($variable)->values()->first());
-                        } else {
-                            $variable = str($variable);
-                            if ($variable->contains('=')) {
-                                // - SESSION_SECRET=123
-                                // - SESSION_SECRET=
-                                $key = $variable->before('=');
-                                $value = $variable->after('=');
-                            } else {
-                                // - SESSION_SECRET
-                                $key = $variable;
-                                $value = null;
-                            }
-                        }
-                    } else {
-                        // SESSION_SECRET: 123
-                        // SESSION_SECRET:
-                        $key = str($variableName);
-                        $value = str($variable);
-                    }
-                    if ($key->startsWith('SERVICE_FQDN')) {
-                        if ($isNew || $savedService->fqdn === null) {
-                            $name = $key->after('SERVICE_FQDN_')->beforeLast('_')->lower();
-                            $fqdn = generateFqdn($resource->server, "{$name->value()}-{$resource->uuid}");
-                            if (substr_count($key->value(), '_') === 3) {
-                                // SERVICE_FQDN_UMAMI_1000
-                                $port = $key->afterLast('_');
-                            } else {
-                                $last = $key->afterLast('_');
-                                if (is_numeric($last->value())) {
-                                    // SERVICE_FQDN_3001
-                                    $port = $last;
-                                } else {
-                                    // SERVICE_FQDN_UMAMI
-                                    $port = null;
-                                }
-                            }
-                            if ($port) {
-                                $fqdn = "$fqdn:$port";
-                            }
-                            if (substr_count($key->value(), '_') >= 2) {
-                                if ($value) {
-                                    $path = $value->value();
-                                } else {
-                                    $path = null;
-                                }
-                                if ($generatedServiceFQDNS->count() > 0) {
-                                    $alreadyGenerated = $generatedServiceFQDNS->has($key->value());
-                                    if ($alreadyGenerated) {
-                                        $fqdn = $generatedServiceFQDNS->get($key->value());
-                                    } else {
-                                        $generatedServiceFQDNS->put($key->value(), $fqdn);
-                                    }
-                                } else {
-                                    $generatedServiceFQDNS->put($key->value(), $fqdn);
-                                }
-                                $fqdn = "$fqdn$path";
-                            }
-
-                            if (! $isDatabase) {
-                                if ($savedService->fqdn) {
-                                    data_set($savedService, 'fqdn', $savedService->fqdn.','.$fqdn);
-                                } else {
-                                    data_set($savedService, 'fqdn', $fqdn);
-                                }
-                                $savedService->save();
-                            }
-                            EnvironmentVariable::create([
-                                'key' => $key,
-                                'value' => $fqdn,
-                                'is_build_time' => false,
-                                'service_id' => $resource->id,
-                                'is_preview' => false,
-                            ]);
-                        }
-                        // Caddy needs exact port in some cases.
-                        if ($predefinedPort && ! $key->endsWith("_{$predefinedPort}")) {
-                            $fqdns_exploded = str($savedService->fqdn)->explode(',');
-                            if ($fqdns_exploded->count() > 1) {
-                                continue;
-                            }
-                            $env = EnvironmentVariable::where([
-                                'key' => $key,
-                                'service_id' => $resource->id,
-                            ])->first();
-                            if ($env) {
-                                $env_url = Url::fromString($savedService->fqdn);
-                                $env_port = $env_url->getPort();
-                                if ($env_port !== $predefinedPort) {
-                                    $env_url = $env_url->withPort($predefinedPort);
-                                    $savedService->fqdn = $env_url->__toString();
-                                    $savedService->save();
-                                }
-                            }
-                        }
-
-                        // data_forget($service, "environment.$variableName");
-                        // $yaml = data_forget($yaml, "services.$serviceName.environment.$variableName");
-                        // if (count(data_get($yaml, 'services.' . $serviceName . '.environment')) === 0) {
-                        //     $yaml = data_forget($yaml, "services.$serviceName.environment");
-                        // }
-                        continue;
-                    }
-                    if ($value?->startsWith('$')) {
-                        $foundEnv = EnvironmentVariable::where([
-                            'key' => $key,
-                            'service_id' => $resource->id,
-                        ])->first();
-                        $value = str(replaceVariables($value));
-                        $key = $value;
-                        if ($value->startsWith('SERVICE_')) {
-                            $foundEnv = EnvironmentVariable::where([
-                                'key' => $key,
-                                'service_id' => $resource->id,
-                            ])->first();
-                            ['command' => $command, 'forService' => $forService, 'generatedValue' => $generatedValue, 'port' => $port] = parseEnvVariable($value);
-                            if (! is_null($command)) {
-                                if ($command?->value() === 'FQDN' || $command?->value() === 'URL') {
-                                    if (Str::lower($forService) === $serviceName) {
-                                        $fqdn = generateFqdn($resource->server, $containerName);
-                                    } else {
-                                        $fqdn = generateFqdn($resource->server, Str::lower($forService).'-'.$resource->uuid);
-                                    }
-                                    if ($port) {
-                                        $fqdn = "$fqdn:$port";
-                                    }
-                                    if ($foundEnv) {
-                                        $fqdn = data_get($foundEnv, 'value');
-                                        // if ($savedService->fqdn) {
-                                        //     $savedServiceFqdn = Url::fromString($savedService->fqdn);
-                                        //     $parsedFqdn = Url::fromString($fqdn);
-                                        //     $savedServicePath = $savedServiceFqdn->getPath();
-                                        //     $parsedFqdnPath = $parsedFqdn->getPath();
-                                        //     if ($savedServicePath != $parsedFqdnPath) {
-                                        //         $fqdn = $parsedFqdn->withPath($savedServicePath)->__toString();
-                                        //         $foundEnv->value = $fqdn;
-                                        //         $foundEnv->save();
-                                        //     }
-                                        // }
-                                    } else {
-                                        if ($command->value() === 'URL') {
-                                            $fqdn = str($fqdn)->after('://')->value();
-                                        }
-                                        EnvironmentVariable::create([
-                                            'key' => $key,
-                                            'value' => $fqdn,
-                                            'is_build_time' => false,
-                                            'service_id' => $resource->id,
-                                            'is_preview' => false,
-                                        ]);
-                                    }
-                                    if (! $isDatabase) {
-                                        if ($command->value() === 'FQDN' && is_null($savedService->fqdn) && ! $foundEnv) {
-                                            $savedService->fqdn = $fqdn;
-                                            $savedService->save();
-                                        }
-                                        // Caddy needs exact port in some cases.
-                                        if ($predefinedPort && ! $key->endsWith("_{$predefinedPort}") && $command?->value() === 'FQDN' && $resource->server->proxyType() === 'CADDY') {
-                                            $fqdns_exploded = str($savedService->fqdn)->explode(',');
-                                            if ($fqdns_exploded->count() > 1) {
-                                                continue;
-                                            }
-                                            $env = EnvironmentVariable::where([
-                                                'key' => $key,
-                                                'service_id' => $resource->id,
-                                            ])->first();
-                                            if ($env) {
-                                                $env_url = Url::fromString($env->value);
-                                                $env_port = $env_url->getPort();
-                                                if ($env_port !== $predefinedPort) {
-                                                    $env_url = $env_url->withPort($predefinedPort);
-                                                    $savedService->fqdn = $env_url->__toString();
-                                                    $savedService->save();
-                                                }
-                                            }
-                                        }
-                                    }
-                                } else {
-                                    $generatedValue = generateEnvValue($command, $resource);
-                                    if (! $foundEnv) {
-                                        EnvironmentVariable::create([
-                                            'key' => $key,
-                                            'value' => $generatedValue,
-                                            'is_build_time' => false,
-                                            'service_id' => $resource->id,
-                                            'is_preview' => false,
-                                        ]);
-                                    }
-                                }
-                            }
-                        } else {
-                            if ($value->contains(':-')) {
-                                $key = $value->before(':');
-                                $defaultValue = $value->after(':-');
-                            } elseif ($value->contains('-')) {
-                                $key = $value->before('-');
-                                $defaultValue = $value->after('-');
-                            } elseif ($value->contains(':?')) {
-                                $key = $value->before(':');
-                                $defaultValue = $value->after(':?');
-                            } elseif ($value->contains('?')) {
-                                $key = $value->before('?');
-                                $defaultValue = $value->after('?');
-                            } else {
-                                $key = $value;
-                                $defaultValue = null;
-                            }
-                            $foundEnv = EnvironmentVariable::where([
-                                'key' => $key,
-                                'service_id' => $resource->id,
-                            ])->first();
-                            if ($foundEnv) {
-                                $defaultValue = data_get($foundEnv, 'value');
-                            }
-                            EnvironmentVariable::updateOrCreate([
-                                'key' => $key,
-                                'service_id' => $resource->id,
-                            ], [
-                                'value' => $defaultValue,
-                                'is_build_time' => false,
-                                'service_id' => $resource->id,
-                                'is_preview' => false,
-                            ]);
-                        }
-                    }
-                }
-                // Add labels to the service
-                if ($savedService->serviceType()) {
-                    $fqdns = generateServiceSpecificFqdns($savedService);
-                } else {
-                    $fqdns = collect(data_get($savedService, 'fqdns'))->filter();
-                }
-                $defaultLabels = defaultLabels($resource->id, $containerName, type: 'service', subType: $isDatabase ? 'database' : 'application', subId: $savedService->id);
-                $serviceLabels = $serviceLabels->merge($defaultLabels);
-                if (! $isDatabase && $fqdns->count() > 0) {
-                    if ($fqdns) {
-                        $shouldGenerateLabelsExactly = $resource->server->settings->generate_exact_labels;
-                        if ($shouldGenerateLabelsExactly) {
-                            switch ($resource->server->proxyType()) {
-                                case ProxyTypes::TRAEFIK->value:
-                                    $serviceLabels = $serviceLabels->merge(fqdnLabelsForTraefik(
-                                        uuid: $resource->uuid,
-                                        domains: $fqdns,
-                                        is_force_https_enabled: true,
-                                        serviceLabels: $serviceLabels,
-                                        is_gzip_enabled: $savedService->isGzipEnabled(),
-                                        is_stripprefix_enabled: $savedService->isStripprefixEnabled(),
-                                        service_name: $serviceName,
-                                        image: data_get($service, 'image')
-                                    ));
-                                    break;
-                                case ProxyTypes::CADDY->value:
-                                    $serviceLabels = $serviceLabels->merge(fqdnLabelsForCaddy(
-                                        network: $resource->destination->network,
-                                        uuid: $resource->uuid,
-                                        domains: $fqdns,
-                                        is_force_https_enabled: true,
-                                        serviceLabels: $serviceLabels,
-                                        is_gzip_enabled: $savedService->isGzipEnabled(),
-                                        is_stripprefix_enabled: $savedService->isStripprefixEnabled(),
-                                        service_name: $serviceName,
-                                        image: data_get($service, 'image')
-                                    ));
-                                    break;
-                            }
-                        } else {
-                            $serviceLabels = $serviceLabels->merge(fqdnLabelsForTraefik(
-                                uuid: $resource->uuid,
-                                domains: $fqdns,
-                                is_force_https_enabled: true,
-                                serviceLabels: $serviceLabels,
-                                is_gzip_enabled: $savedService->isGzipEnabled(),
-                                is_stripprefix_enabled: $savedService->isStripprefixEnabled(),
-                                service_name: $serviceName,
-                                image: data_get($service, 'image')
-                            ));
-                            $serviceLabels = $serviceLabels->merge(fqdnLabelsForCaddy(
-                                network: $resource->destination->network,
-                                uuid: $resource->uuid,
-                                domains: $fqdns,
-                                is_force_https_enabled: true,
-                                serviceLabels: $serviceLabels,
-                                is_gzip_enabled: $savedService->isGzipEnabled(),
-                                is_stripprefix_enabled: $savedService->isStripprefixEnabled(),
-                                service_name: $serviceName,
-                                image: data_get($service, 'image')
-                            ));
-                        }
-                    }
-                }
-                if ($resource->server->isLogDrainEnabled() && $savedService->isLogDrainEnabled()) {
-                    data_set($service, 'logging', [
-                        'driver' => 'fluentd',
-                        'options' => [
-                            'fluentd-address' => 'tcp://127.0.0.1:24224',
-                            'fluentd-async' => 'true',
-                            'fluentd-sub-second-precision' => 'true',
-                        ],
-                    ]);
-                }
-                if ($serviceLabels->count() > 0) {
-                    if ($resource->is_container_label_escape_enabled) {
-                        $serviceLabels = $serviceLabels->map(function ($value, $key) {
-                            return escapeDollarSign($value);
-                        });
-                    }
-                }
-                data_set($service, 'labels', $serviceLabels->toArray());
-                data_forget($service, 'is_database');
-                if (! data_get($service, 'restart')) {
-                    data_set($service, 'restart', RESTART_MODE);
-                }
-                if (data_get($service, 'restart') === 'no' || data_get($service, 'exclude_from_hc')) {
-                    $savedService->update(['exclude_from_status' => true]);
-                }
-                data_set($service, 'container_name', $containerName);
-                data_forget($service, 'volumes.*.content');
-                data_forget($service, 'volumes.*.isDirectory');
-                data_forget($service, 'volumes.*.is_directory');
-                data_forget($service, 'exclude_from_hc');
-                data_set($service, 'environment', $serviceVariables->toArray());
-                updateCompose($savedService);
-
-                return $service;
-
-            });
-
-            $envs_from_coolify = $resource->environment_variables()->get();
-            $services = collect($services)->map(function ($service, $serviceName) use ($resource, $envs_from_coolify) {
-                $serviceVariables = collect(data_get($service, 'environment', []));
-                $parsedServiceVariables = collect([]);
-                foreach ($serviceVariables as $key => $value) {
-                    if (is_numeric($key)) {
-                        $value = str($value);
-                        if ($value->contains('=')) {
-                            $key = $value->before('=')->value();
-                            $value = $value->after('=')->value();
-                        } else {
-                            $key = $value->value();
-                            $value = null;
-                        }
-                        $parsedServiceVariables->put($key, $value);
-                    } else {
-                        $parsedServiceVariables->put($key, $value);
-                    }
-                }
-                $parsedServiceVariables->put('COOLIFY_CONTAINER_NAME', "$serviceName-{$resource->uuid}");
-                $parsedServiceVariables = $parsedServiceVariables->map(function ($value, $key) use ($envs_from_coolify) {
-                    if (! str($value)->startsWith('$')) {
-                        $found_env = $envs_from_coolify->where('key', $key)->first();
-                        if ($found_env) {
-                            return $found_env->value;
-                        }
-                    }
-
-                    return $value;
-                });
-
-                data_set($service, 'environment', $parsedServiceVariables->toArray());
-
-                return $service;
-            });
-            $finalServices = [
-                'services' => $services->toArray(),
-                'volumes' => $topLevelVolumes->toArray(),
-                'networks' => $topLevelNetworks->toArray(),
-                'configs' => $topLevelConfigs->toArray(),
-                'secrets' => $topLevelSecrets->toArray(),
-            ];
-            $yaml = data_forget($yaml, 'services.*.volumes.*.content');
-            $resource->docker_compose_raw = Yaml::dump($yaml, 10, 2);
-            $resource->docker_compose = Yaml::dump($finalServices, 10, 2);
-            $resource->save();
-            $resource->saveComposeConfigs();
-
-            return collect($finalServices);
-        } else {
-            return collect([]);
-        }
-    }
 }
 
 function parseEnvVariable(Str|string $value)
