+++ resources/views/livewire/project/service/storage.blade.php
@@ -31,10 +31,13 @@
         @endif
 
         @if ($resource->persistentStorages()->get()->count() > 0)
+            <h3 class="pt-4">Volumes</h3>
             <livewire:project.shared.storages.all :resource="$resource" />
         @endif
         @if ($fileStorage->count() > 0)
-            <div class="flex flex-col gap-4 pt-4">
+
+            <h3 class="mt-4 pt-2  border-t dark:border-coolgray-200">Mounts</h3>
+            <div class="flex flex-col gap-2">
                 @foreach ($fileStorage->sort() as $fileStorage)
                     <livewire:project.service.file-storage :fileStorage="$fileStorage"
                         wire:key="resource-{{ $fileStorage->uuid }}" />
